misc/HtmlTestSuite/results/test-0451.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8"/>
<title>CSS Test: line-height resolved value</title>
<link href="https://drafts.csswg.org/cssom/#resolved-values" rel="help"/>
<link href="https://github.com/w3c/csswg-drafts/issues/3749" rel="help"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<style>
  div { font-size: 16px; }
</style>
</head><body><div data-expected="normal" style="line-height: normal"/>
<div data-expected="16px" style="line-height: 1"/>
<div data-expected="10px" style="line-height: 10px"/>
<div data-expected="1.6px" style="line-height: 10%"/>
<script>
for (const e of document.querySelectorAll("div")) {
  const specified = e.style.lineHeight;
  test(function() {
    const expected = e.getAttribute("data-expected");
    assert_equals(getComputedStyle(e).lineHeight, expected, specified + " should compute to " + expected);
  }, "line-height: " + specified);
}
</script>
</body></html>
